Change approach to intercept all clicks
Change the approach to capture clicks at the root
element and then determine if it is a click on a
book icon. The major improvement here, is that
images always work, and there is not wait for the
page the load + running a script to attach callbacks
to all the images.